Backend/MiddleWare/requirements.txt
flask==2.3.2
mysql-connector-python==8.0.33
requests==2.28.1
bcrypt==4.0.1
flask-cors